+++ apps/fabric-website-resources/src/components/pages/CoachmarkPage.tsx
@@ -4,5 +4,8 @@ import { DemoPage } from '../DemoPage';
 import { CoachmarkPageProps } from 'office-ui-fabric-react/lib/components/Coachmark/Coachmark.doc';
 
 export const CoachmarkPage = (props: { isHeaderVisible: boolean }) => (
-  <DemoPage {...{ ...CoachmarkPageProps, ...props }} />
+  <DemoPage
+    jsonDocs={require('@uifabric/api-docs/lib/pages/office-ui-fabric-react/Coachmark.page.json')}
+    {...{ ...CoachmarkPageProps, ...props }}
+  />
 );