+++ apps/fabric-website/src/styles/constants.ts
@@ -7,8 +7,3 @@ export const contentPaddingBottom = 100;
 
 export const queryUhfMobileMin = `@media screen and (min-width: ${ScreenWidthMinUhfMobile}px)`;
 export const queryXLargeMin = `@media screen and (min-width: ${ScreenWidthMinXLarge}px)`;
-
-export const uhfMarginH3 = '40px 0 8px 0';
-export const uhfMarginH2 = '48px 0 4px 0';
-export const uhfMarginH4 = '8px 0';
-export const uhfMarginBottomP = 12;