+++ packages/example-app-base/src/components/Markdown/index.ts
@@ -1,0 +1,5 @@
+export * from './MarkdownParagraph';
+export * from './MarkdownCode';
+export * from './Markdown';
+export * from './Markdown.types';
+export * from './MarkdownHeader';