+++ packages/example-app-base/src/utilities/extractAnchorLink.ts
@@ -1,0 +1,10 @@
+/**
+ * Returns the last hash value in the link
+ */
+export function extractAnchorLink(url: string): string | undefined {
+  const split = url.split('#');
+  const cleanedSplit = split.filter((value: string) => !!value);
+  if (cleanedSplit.length > 1) {
+    return cleanedSplit[cleanedSplit.length - 1];
+  }
+}