+++ packages/office-ui-fabric-react/src/components/Coachmark/Coachmark.doc.tsx
@@ -16,7 +16,6 @@ export const CoachmarkPageProps: IDocPageProps = {
       isScrollable: false
     }
   ],
-  propertiesTablesSources: [require<string>('!raw-loader!office-ui-fabric-react/src/components/Coachmark/Coachmark.types.ts')],
   overview: require<string>('!raw-loader!office-ui-fabric-react/src/components/Coachmark/docs/CoachmarkOverview.md'),
   bestPractices: '',
   dos: require<string>('!raw-loader!office-ui-fabric-react/src/components/Coachmark/docs/CoachmarkDos.md'),