+++ apps/fabric-website-resources/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @uifabric/fabric-website-resources
 
-This log was last generated on Wed, 24 Apr 2019 12:35:54 GMT and should not be manually modified.
+This log was last generated on Tue, 30 Apr 2019 12:31:36 GMT and should not be manually modified.
+
+## 6.15.0
+Tue, 30 Apr 2019 12:31:36 GMT
+
+### Minor changes
+
+- change component pages to use page json files
 
 ## 6.14.4
 Wed, 24 Apr 2019 12:35:54 GMT